Split CacheStatus into its own source file
